+++ app/components/SignUp.jsx
@@ -18,8 +18,6 @@ const style = {
   }
 };
 
-
-
 function mapDispatchToProps(dispatch) {
   return {
     signUpSubmit: function(name, email, password) {
